Update related-standards.md
Now only the standards that are used in the mapping table are discussed here in this file. The additional standards are commented and open for review, if we should skip them or if we would like to also map then to the OWASP OT Top 10 items in the mapping table.